Add minimum spanning tree generation to AllocatorGraphs
git-svn-id: file:///A/Documents/Unreal%20Projects/RogueSkySVNtoGit@15 12684d69-bc46-6143-8e79-893cf67d636b

Former-commit-id: 69fbb9062c68599b3015818f3aa13c4315fe8366

+++ Source/RogueSky/Public/Generation/GenerationConstants.h
@@ -62,7 +62,7 @@ namespace GenerationConstants {
     const uint8 MAX_CHUNK_RESOLUTION = 32;
     const uint16 VOXEL_SIZE = CHUNK_SIZE / MAX_CHUNK_RESOLUTION;
     const uint8 DISTANCE_FIELD_SIZE = MAX_CHUNK_RESOLUTION + 1;
-    const uint8 MAX_LOD_DEPTH = FMath::Log2(MAX_CHUNK_RESOLUTION);
+    const uint8 MAX_LOD_DEPTH = FMath::Log2((uint8)MAX_CHUNK_RESOLUTION);
 
     static FVector ChunkCoordinateToWorld(int ChunkX, int ChunkY, int ChunkZ, int X, int Y, int Z) {
         return FVector(ChunkX * CHUNK_SIZE + X * VOXEL_SIZE, ChunkY * CHUNK_SIZE + Y * VOXEL_SIZE, ChunkZ * CHUNK_SIZE + Z * VOXEL_SIZE);
